FVI refactoring - moved feature comparer into nodes.
Means nodes can store their children in an ordered fashion - which can be leveraged by the index itself for performance. makes more sense to have the nodes responsible for this anyway, since for nodes backed by external storage, that external storage is going to have the nodes structured in accordance with a particular ordering.

As part of this, added a FeatureVectorComponent struct type (instead of using KeyValuePairs), and slightly changed the node interfaces to make the async and non-async versions more similar.